- Updated ChangeLog - Added "Elves Wisdom Bonus" that controls the amount of bonus in Mana for Elves. Set to 1.0 to disable the Racial feature by calculations.
With this, I finished a basic support for Elven Features (Just basic because I dont have the real numbers for bonuses and features, just a explanation how they works) BUT, there are two Elven features that I cant implement. "Infused with Magic" because I cant see the Cap for energy resistance (Maybe who worked resistances can implement this better hehe) and "Perception", because I cant find the Passive Detect Hidden (Its exists in WP?) to work on the bonuses.