Two additions to onSeeChar and onSeeItem events. Now in onSeeChar source char is not counted and invisible chars are ignored (NPC Cant see them, right?) On onSeeItem invisible items are ignored.
In both events, python event will just be called if NPC is dont nothing at this time. So, in combat or other actions, this event will not be called.

Good way to customize a lot of things in AI by python, like Wanted items, natural enemies, reaction seeing chars, etc etc.
Just use carefully. If we have lots of npcs with these triggers, performance could be decreased a little.